Big changes TODO AMMEND ME

+++ app/controller_donate.mjs
@@ -2,21 +2,21 @@ import assert from 'assert';
 import _ from 'lodash';
 
 import * as db from"./db.mjs";
-import { getRemoteIp, hashCode, projectRows } from "./util.mjs";
+import { foldListInPlace, getRemoteIp, hashCode, projectList, projectListInPlace } from "./util.mjs";
 import { GIVE_FULL_LEASE_SET, LEASE_MAX_AGE_SECS, LEASE_SET_SIZE, MAX_AGE_SECS } from './config.mjs';
 
 // turns [{crypto_symbol: 'XXX', a: A ...}, ...]
 // into  {XXX: [{crypto_symbol: 'XXX', a: A ...}, ...], YYY: ...}
 // easier to grok for templates.
-function foldLeaseSet(leases) {
-  const rv = {};
-  _.forEach(leases, (lease) => {
-    rv[lease.crypto_symbol] = rv[lease.crypto_symbol] || [];
-    rv[lease.crypto_symbol].push(lease);
-  });
+// function foldLeaseSet(leases) {
+//   const rv = {};
+//   _.forEach(leases, (lease) => {
+//     rv[lease.crypto_symbol] = rv[lease.crypto_symbol] || [];
+//     rv[lease.crypto_symbol].push(lease);
+//   });
 
-  return rv;
-}
+//   return rv;
+// }
 
 function leaseSetPickOneInPlace({ leaseSet, ip, userAgent }) {
   assert(typeof(leaseSet) === 'object'); // quick check it is folded.
@@ -34,35 +34,38 @@ function leaseSetPickOneInPlace({ leaseSet, ip, userAgent }) {
 }
 
 export function mw_handle_get_donate(req, res, next) {
-  const ip = getRemoteIp(req);
+  const ip = getRemoteIp(req).trim();
   const userAgent = req.headers['user-agent'] || null;
+  if (!ip) throw new Error();
   
   db.beginTransaction(req);
 
   db.destroyExpiredLeases({ ip });
-  let leaseSet = db.getLeases({ ip });
+  let leaseSet = db.getLeases({ filter: { ip } });
 
   if (leaseSet.length === 0) {
-    // first time call or previous leases expired, get new ones for this ip.
-    const cryptoSymbols = db.getAllCryptoSymbols();
+    // First time for this ip, or previous leases expired.
+    // For each coin, get fresh leases and also insert them in db.
 
+    const cryptoSymbols = db.getAllCryptoSymbols();
     leaseSet = [];
     _.forEach(cryptoSymbols, (cryptoSymbol) => {
       const newLeasesForSymbol = db.getFreshLeases({ limit: LEASE_SET_SIZE, cryptoSymbol });
       db.insertLeases({ cryptoSymbol, ip, userAgent, leases: newLeasesForSymbol });
+
       leaseSet.push(newLeasesForSymbol);
     });
-    db.commitTransaction(req);
 
     leaseSet = _.flatten(leaseSet);
-  } else {
-    db.commitTransaction(req);
   }
+
+  db.commitTransaction(req);
   
-  leaseSet = foldLeaseSet(projectRows(leaseSet, ['address', 'crypto_symbol']));
+  // XXX rename
+  projectListInPlace(leaseSet, ['pool_rowid.address', 'pool_rowid.crypto_symbol']);
 
   if (!GIVE_FULL_LEASE_SET) {
-    leaseSetPickOneInPlace({ leaseSet, ip, userAgent });
+    //leaseSetPickOneInPlace({ leaseSet, ip, userAgent });
   }
 
   res.data.leaseSet = leaseSet;
